requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.8
requests==2.32.5

numpy==1.26.4
pandas==2.3.2
scikit-learn==1.7.2
scipy==1.15.3

transformers==4.44.2
tokenizers==0.19.1
sentencepiece==0.2.1
sentence-transformers==5.1.0
huggingface_hub==0.34.0
accelerate==1.0.1

faiss-cpu==1.12.0

nltk==3.9.1
bert-score==0.3.13
rouge-score==0.1.2

python-dotenv==1.0.0
uvloop==0.17.0
gradio==5.45.0